healthbar ready to implement
