--- maui-and-the-sun/Assets/MainMenu.cs
@@ -1,19 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class MainMenu : MonoBehaviour
-{
-    public void PlayGame()
-    {
-        SceneManager.LoadScene(SceneManager. GetActiveScene().buildIndex + 1);
-
-    }
-    public void QuiteGame()
-    {
-        Debug.Log("Quite!");
-        Application.Quit();
-       
-    }
-}
