+++ maui-and-the-sun/Assets/Project- Bishesh/Scripts/HealthBar.cs
@@ -1,0 +1,19 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.UI;
+
+public class HealthBar : MonoBehaviour
+{
+    public Slider slider;
+    
+    public void setMaxHealth (int health)
+    {
+        slider.maxValue = health;
+        slider.value = health;
+    }
+    public void setHealth (int health)
+    {
+        slider.value = health;
+    }
+}
