+++ maui-and-the-sun/Library/StateCache/SceneView/42a6dd-mainStage.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":true,"pivot":{"x":285.87969970703127,"y":180.20054626464845,"z":-2335.66748046875},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":461.63787841796877,"orthographic":true}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"showFog":true,"showMaterialUpdate":false,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true},"in2DMode":true,"pivot":{"x":9.029019355773926,"y":5.082974433898926,"z":-3147.65625},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":28.13141632080078,"orthographic":true}
